+++ samples/Agar.Godot/Client/Scripts/Networking/AgarNetworkSession.cs
@@ -4,6 +4,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Rpc;
+using Rpc.Generated;
 using Shared.Interfaces;
 using ULinkRPC.Client;