+++ samples/Agar.Unity/Client/Assets/Scripts/Gameplay/DotArenaNetworkSession.cs
@@ -4,7 +4,9 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using Rpc;
+using Rpc.Generated;
 using Shared.Interfaces;
+using ULinkRPC.Client;
 
 namespace SampleClient.Gameplay
 {