+++ samples/Agar.Unity/Client/Assets/Scripts/Rpc/Testing/RpcConnectionTester.cs
@@ -3,6 +3,7 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Rpc.Generated;
 using Shared.Interfaces;
 using ULinkRPC.Client;
 using UnityEngine;